+++ zsh/.config/zsh/alias.zsh
@@ -1,0 +1,10 @@
+# Terminal aliases
+alias c='clear'
+alias e='exit'
+alias sz='source ~/.zshrc'
+
+# Apps
+alias cat='bat'
+alias ls='eza'
+alias l='eza -l'
+alias ll='eza -lah'