+++ zsh/.config/zsh/plugins.zsh
@@ -1,0 +1,8 @@
+plugins=(
+    fzf
+    git
+    vi-mode
+    zoxide
+    zsh-autosuggestions
+    zsh-syntax-highlighting
+)